requirements.txt
# for adding new packages, add a locked version here and then run `uv pip install -r requirements.txt`
python-dotenv==1.1.0

# Streamlit release notes: https://docs.streamlit.io/develop/quick-reference/release-notes
streamlit==1.45.1
authlib>=1.6.0  # required for Streamlit authentication
pandas==2.2.3  # for data manipulation and Streamlit tables

# Database
sqlalchemy==2.0.41
psycopg2-binary==2.9.10

# Optional packages
# litellm==1.63.14 # to simplify model-provider routing
# openai==1.30.1  # optional, for chat UI demo purposes
